LayoutTests/http/tests/media/fairplay/fps-video-frame-expected.txt


EVENT(encrypted)
EVENT(message)
PROMISE: licenseResponse resolved
EVENT(canplay)
isBlack is true
END OF TEST
